Update Product & Product Variation to use Double in stock quantity.

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/ProductVariation.kt
@@ -35,7 +35,7 @@ data class ProductVariation(
     val isSaleScheduled: Boolean,
     val stockStatus: ProductStockStatus,
     val backorderStatus: ProductBackorderStatus,
-    val stockQuantity: Int,
+    val stockQuantity: Double,
     val options: List<Option>,
     var priceWithCurrency: String? = null,
     val isPurchasable: Boolean,
